+++ src/processors/code-block.ts
@@ -2,10 +2,8 @@ import LocalQuotes from "../main";
 import {BlockMetadata, selectBlockMetadata} from "../types/block-metadata";
 import {getAuthorsCode, updateQuotesVault} from "../types/quote";
 import {findTaggedFiles} from "../utils/scan";
-import {parseMdToHtml} from "../utils/parser";
 import {OneTimeBlock, selectOneTimeBlock} from "../types/one-time-block";
-import {MarkdownPostProcessorContext} from "obsidian";
-import {isStringWeakForInnerHtmlVulnerability} from "../utils/dom";
+import { MarkdownPostProcessorContext, MarkdownRenderer } from 'obsidian';
 
 export async function processCodeBlock(
 	plugin: LocalQuotes,
@@ -23,19 +21,13 @@ export async function processCodeBlock(
 	const bq: HTMLElement = el.createEl('blockquote');
 	el.appendChild(bq);
 
-	if (isStringWeakForInnerHtmlVulnerability(blockMetadata.content.text + blockMetadata.content.author)) {
-		bq.innerText = 'Remove script tag from your quote, tricky fox!';
-		return;
-	}
-
 	for (let p of plugin.settings.quoteBlockFormat.split('\n')) {
-		bq.innerHTML += parseMdToHtml(
+		await MarkdownRenderer.renderMarkdown(
 			p.replace('{{content}}', blockMetadata.content.text.split('\n').join('<br/>'))
-			.replace('{{author}}',
-				plugin.settings.inheritListingStyle
-					? getAuthorsCode(plugin.settings.quoteVault, blockMetadata.content.author)
-					: blockMetadata.content.author
-			)
+			.replace('{{author}}', plugin.settings.inheritListingStyle
+				? getAuthorsCode(plugin.settings.quoteVault, blockMetadata.content.author)
+				: blockMetadata.content.author),
+			bq, null, null
 		);
 	}
 }
@@ -55,19 +47,13 @@ export async function processOneTimeCodeBlock(
 	const bq: HTMLElement = el.createEl('blockquote');
 	el.appendChild(bq);
 
-	if (isStringWeakForInnerHtmlVulnerability(oneTimeBlock.content.text + oneTimeBlock.content.author)) {
-		bq.innerText = 'Remove script tag from your quote, tricky fox!';
-		return;
-	}
-
 	for (let p of plugin.settings.quoteBlockFormat.split('\n')) {
-		bq.innerHTML += parseMdToHtml(
+		await MarkdownRenderer.renderMarkdown(
 			p.replace('{{content}}', oneTimeBlock.content.text.split('\n').join('<br/>'))
-				.replace('{{author}}',
-					plugin.settings.inheritListingStyle
-						? getAuthorsCode(plugin.settings.quoteVault, oneTimeBlock.content.author)
-						: oneTimeBlock.content.author
-				)
+			.replace('{{author}}', plugin.settings.inheritListingStyle
+				? getAuthorsCode(plugin.settings.quoteVault, oneTimeBlock.content.author)
+				: oneTimeBlock.content.author),
+			bq, null, null
 		);
 	}
 }