+++ src/utils/scan.ts
@@ -1,4 +1,4 @@
-import {TFile} from "obsidian";
+import { TFile } from 'obsidian';
 import LocalQuotes from "../main";
 import {Quote} from "../types/quote";
 
@@ -10,7 +10,7 @@ export function getBlockMetadataIdx(plugin: LocalQuotes, id: string): number {
 	return plugin.settings.blockMetadata.findIndex((e) => e.id === id);
 }
 
-function checkFileTag(f: TFile, tag: string): boolean {
+export function checkFileTag(f: TFile, tag: string): boolean {
 	const tagInContent = app.metadataCache.getFileCache(f).tags &&
 		(app.metadataCache.getFileCache(f).tags.findIndex((t) => t.tag === `#${tag}`) >= 0);
 	const tagInFrontmatter = app.metadataCache.getFileCache(f).frontmatter &&