+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Some fixes from [here](https://github.com/ka1tzyu/local-quotes/projects/1#card-82152799)
 
 ## [1.2.9] - 2022/05/13
 ### Fixed