+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Replace showdown to Obsidian's `renderMarkdown()`, so plugin size reduced to ~35KB 
+([related task](https://github.com/ka1tzyu/local-quotes/projects/1#card-82152964))
+
 ### Fixed
 - Some fixes from [here](https://github.com/ka1tzyu/local-quotes/projects/1#card-82152799)