+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## Added
+- Ability to automatically update quote listing when file modifies, so you don't need to manually rescan it
+- Setting for automatically quote listing update (enabled by default) 
 
 ## [1.2.10] - 2022/05/19
 ### Changed